code/requirements.txt
requests
docker
nuvla-api